products/prompt-lock/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "prompt-lock"
version = "0.1.0"
description = "Git-native prompt regression testing with judge calibration for LLM CI/CD pipelines"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
keywords = ["llm", "testing", "ci-cd", "prompt", "evaluation", "regression", "judge-calibration"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Testing",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "click>=8.1",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "gitpython>=3.1",
    "litellm>=1.0",
    "sentence-transformers>=5.0",
    "rich>=13.0",
    "scipy>=1.11",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.0",
    "black>=24.0",
    "ruff>=0.3",
    "mypy>=1.8",
]

[project.scripts]
prompt-lock = "prompt_lock.cli:main"

[project.urls]
Homepage = "https://github.com/buildworld-ai/prompt-lock"
Repository = "https://github.com/buildworld-ai/prompt-lock"
Issues = "https://github.com/buildworld-ai/prompt-lock/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/prompt_lock"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.ruff]
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.mypy]
python_version = "3.11"
strict = false
ignore_missing_imports = true

[tool.coverage.run]
source = ["src/prompt_lock"]
